components/auto-labeler/requirements.txt
google-cloud-storage
opencv-python-headless
mediapipe